assets/pyproject.toml
[project]
name = "assets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycairo==1.27.0",
    "pydantic>=2.11.7",
    "python-frontmatter==1.1.0",
    "PyYAML==6.0.2",
]
